refactor pt 6: split device_step.py into config/state/kernels
- split jax_ops/device_step.py (1273 LOC) into device_config (host DeviceConfig + lap-plan builders), device_state (host<->device marshalling), device_kernels (traced substeps + compiled step + run_window host loop)
- device_step keeps its design docstring, becomes a re-export shim so device_step.run_window is unchanged
- no arithmetic moved: jax-device bit-identical to pre-split (2D/diffusion/3D), gate green, hybrid jax unaffected
